refactor: Use css module in options components
