+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
   "editor.codeActionsOnSave": {
     "source.fixAll": "explicit"
-  }
+  },
+  "css.lint.validProperties": ["composes"]
 }
