--- src/components/options/LanguageOptions/LanguageOptions.styles.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components'
-
-export const Flag = styled.img`
-  width: 1.1rem;
-  border-radius: 4px;
-`
-
-export const OptionItemFlag = styled(Flag)`
-  margin-right: 0.5rem;
-`
