+++ src/components/options/Option/Option.module.css
@@ -1,0 +1,17 @@
+.optionContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: center;
+}
+
+.optionSelection {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  margin-left: 12px;
+}
+
+.optionIcon {
+  margin-right: 0.5rem;
+}
