--- src/components/options/Option/Option.styles.ts
@@ -1,21 +1,0 @@
-import styled from 'styled-components'
-
-export const OptionContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: center;
-`
-
-export const OptionName = styled.strong``
-
-export const OptionSelection = styled.span`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  margin-left: 12px;
-`
-
-export const OptionIcon = styled.span`
-  margin-right: 0.5rem;
-`
