+++ src/components/options/OptionItem/OptionItem.module.css
@@ -1,0 +1,10 @@
+.optionItem {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  margin-left: 12px;
+}
+
+.optionIcon {
+  margin-right: 0.5rem;
+}
