--- src/components/options/ThemeOptions/ThemeOptions.styles.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-
-export const OptionItem = styled.span`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  margin-left: 12px;
-`
-
-export const OptionIcon = styled.svg`
-  margin-right: 0.5rem;
-`
